sensors: Bulk update.
Changes:
 * Add new variable GLASS_MULTIPLIER.
 * Use uint16_t instead of size_t on ADC value passing.
 * Modify proximity sensor values to adapt GLASS_MULTIPLIER.
 * Modify tabs to space and fix indentation.
 * Remove range configuring on Proximity sensor as it's not needed
   there.
 * Modify light sensor resolution to 1.0f since roundf converts it.

+++ libsensors/LightSensor.h
@@ -38,9 +38,9 @@ private:
     bool mHasPendingEvent;
     sensors_event_t mPendingEvent;
 
-    float getLuxValue(size_t adc_count);
-    float getCorrectADC(size_t adc_count);
-    void configureRange(size_t adc_count);
+    float getLuxValue(uint16_t adc_count);
+    float getCorrectADC(uint16_t adc_count);
+    void configureRange(uint16_t adc_count);
     struct aps_12d_settings settings;
 public:
             LightSensor();
